requirements.txt
matplotlib
numpy
pyelftools
pyserial
tabulate
cfclient